+++ lib/geniuslink_design_system.dart
@@ -8,6 +8,7 @@ library geniuslink_design_system;
 
 export 'design_system/tokens.dart';
 export 'design_system/app_theme.dart';
+export 'package:smart_auto_suggest_box/smart_auto_suggest_box.dart' show SmartAutoSuggestBoxLocalizations, SmartAutoSuggestTheme;
 
 export 'design_system/components/core/core_components.dart';
 export 'design_system/components/domain/domain_components.dart';